+++ src/app/change-log/layout.tsx
@@ -1,4 +1,4 @@
-import { MainMenu } from "@/features/menu/menu";
+import { MainMenu } from "@/features/main-menu/menu";
 import { AI_NAME } from "@/features/theme/customise";
 
 export const dynamic = "force-dynamic";
