+++ src/app/unauthorized/layout.tsx
@@ -1,4 +1,4 @@
-import { MainMenu } from "@/features/menu/menu";
+import { MainMenu } from "@/features/main-menu/menu";
 import { AI_NAME } from "@/features/theme/customise";
 
 export const metadata = {
@@ -14,9 +14,7 @@ export default async function RootLayout({
   return (
     <>
       <MainMenu />
-      <div className="flex-1">
-        {children}
-      </div>
+      <div className="flex-1">{children}</div>
     </>
   );
 }
