+++ src/features/chat/chat-services/chat-thread-service.ts
@@ -3,10 +3,10 @@ import "server-only";
 
 import { userHashedId, userSession } from "@/features/auth/helpers";
 import { FindAllChats } from "@/features/chat/chat-services/chat-service";
-import { deleteDocuments } from "@/features/langchain/vector-stores/azure-cog-search/azure-cog-vector-store";
 import { SqlQuerySpec } from "@azure/cosmos";
 import { nanoid } from "nanoid";
 import { CosmosDBContainer } from "../../common/cosmos";
+import { deleteDocuments } from "../vector-stores/azure-cog-search/azure-cog-vector-store";
 import { FindAllChatDocuments } from "./chat-document-service";
 import {
   CHAT_THREAD_ATTRIBUTE,
