+++ src/features/chat/vector-stores/azure-cog-search/azure-cog-vector-store.ts
