+++ src/features/providers.tsx
@@ -2,7 +2,7 @@
 
 import { SessionProvider } from "next-auth/react";
 import { GlobalMessageProvider } from "./global-message/global-message-context";
-import { MenuProvider } from "./menu/menu-context";
+import { MenuProvider } from "./main-menu/menu-context";
 
 export const Providers = ({ children }: { children: React.ReactNode }) => {
   return (
